Completed git practice assignment
I think that this assignment was helpful in furthering my knowledge of git.
At first, I was among those who were a bit confused about which specific deliverables
were required to be completed, but Chris was helpful in clearing that up on Slack.
Going forward, maybe it would be helpful to quickly run through the specifics of each night's
assignment together as a class to prevent similar confusion in the future.  4/5.